+++ tests/Unit/Internal/ContextStub.php
@@ -25,7 +25,7 @@ class ContextStub extends BasicContextStub implements ContextInterface
     private $isAdmin;
     private $skipLogTags;
     private $adminUserId;
-    private $productiveMode;
+    private bool $productiveMode;
     private $demoMode;
 
     /**
@@ -44,6 +44,7 @@ class ContextStub extends BasicContextStub implements ContextInterface
         $this->isAdmin = $context->isAdmin();
         $this->skipLogTags = $context->getSkipLogTags();
         $this->demoMode = $context->isShopInDemoMode();
+        $this->productiveMode = $context->isShopInProductiveMode();
     }
 
     /**
@@ -214,18 +215,12 @@ class ContextStub extends BasicContextStub implements ContextInterface
         $this->skipLogTags = $skipLogTags;
     }
 
-    /**
-     * @return bool
-     */
     public function isShopInProductiveMode(): bool
     {
         return $this->productiveMode;
     }
 
-    /**
-     * @param bool $productiveMode
-     */
-    public function setShopInProductiveMode(bool $productiveMode)
+    public function setShopInProductiveMode(bool $productiveMode): void
     {
         $this->productiveMode = $productiveMode;
     }